+++ Tools/Internal/MyBoxStartupAsset.cs
@@ -1,6 +1,0 @@
-﻿using UnityEngine;
-
-namespace MyBox.Internal
-{
-	public class MyBoxStartupAsset : ScriptableObject { }
-}